142 Linked List CycleII step03まで完了
