--- 141. Linked List Cycle/step03.py
@@ -1,10 +1,0 @@
-class Solution:
-    def hasCycle(self, head: Optional[ListNode]) -> bool:
-        visited_nodes = set()
-        node = head
-        while node:
-            if node in visited_nodes:
-                return True
-            visited_nodes.add(node)
-            node = node.next
-        return False
